test: extract mount request suite
